Move sspi_impl to seperate file
The sspi functionality is an implementation detail so move it to a
seperate file in the detail namespace.
